+++ lib/src/modules/instabug.dart
@@ -412,4 +412,13 @@ class Instabug {
     final darkKey = await dark.obtainKey(configuration);
     return _host.setCustomBrandingImage(lightKey.name, darkKey.name);
   }
+
+  /// Allows detection of app review sessions which are submitted through custom prompts.
+  ///
+  /// Use this when utilizing a custom app rating prompt. It should be called
+  /// once the user clicks on the Call to Action (CTA) that redirects them to the app store.
+  /// Helps track session data for insights on user interactions during review submission.
+  static Future<void> willRedirectToStore() async {
+    return _host.willRedirectToStore();
+  }
 }